Break down v1.Symbol.{Info,Kind,...} into separate top-levels
